+++ doc/changelog.rst
@@ -3,6 +3,7 @@ Changelog
 * :feature:`979` Change Ordered set implementation from orderedset to ordered-set, as the former is not maintained anymore
 * :feature:`598` Drop support for python version < 3.6
 * :feature:`-` Added support for tagging test parameterization values
+* :feature:`-` Added a plugin which provides links to test logs which have been archived by a CI system
 * :release:`1.8.0 <03-07-2019>`
 * :feature:`945` Drop support for deprecated arguments of ``add_cleanup``
 * :feature:`452` Drop support for old-style assertions